tc358840: fix pixelclock computation
There is a possible overflow in the current computation. Here is what it
looked like with a 1080p60 input before the patch:
tc358840 2-001f: event: resolution change (EQ_BYPS=1): 1920x1080p8.17
(2200x1125)

Although the interval received was 16.6ms (or 60.24fps).

After the patch, the pixelclock is properly set with the same setup:
tc358840 2-001f: event: resolution change (EQ_BYPS=1): 1920x1080p60.24
(2200x1125)

+++ drivers/media/i2c/tc358840.c
@@ -1308,7 +1308,7 @@ static int tc358840_get_detected_timings(struct v4l2_subdev *sd,
 	bt->height = height;
 	bt->vsync = frame_height - height;
 	bt->hsync = frame_width - width;
-	bt->pixelclock = DIV_ROUND_CLOSEST((frame_width * frame_height * fps)
+	bt->pixelclock = DIV_ROUND_CLOSEST(((u64)frame_width * frame_height * fps)
 					   / 100, 1000) * 1000;
 
 	if (pol & MASK_S_V_HPOL)
